added inner classes
